requirements.txt


# geopy==2.4.1
# fastapi==0.115.4
# uvicorn==0.32.0
# SQLAlchemy==2.0.36
# psycopg2-binary>=2.9.9
# psycopg>=3.1
# pydantic==2.9.2
# python-dotenv==1.0.1
# pythonping==1.1.4
# datadog-api-client==2.21.0


# --- Core Web Framework ---
fastapi==0.115.4
uvicorn==0.32.0

# --- Database and ORM ---
SQLAlchemy==2.0.36
psycopg2-binary>=2.9.9
psycopg>=3.1
python-dotenv==1.0.1

# --- Network and Monitoring ---
pythonping==1.1.4
datadog-api-client==2.21.0
geopy==2.4.1

# --- Data and Validation ---
pydantic==2.9.2

# --- LangChain + Ollama integration ---
langchain==0.3.7
langchain-core==0.3.18
langchain-community==0.3.6
langchain-ollama==0.1.3
ollama==0.3.3

# --- YAML + JSON Utilities ---
PyYAML==6.0.2

# --- Optional (for safety and compatibility) ---
typing-extensions>=4.9.0
